Have maturin use $PYTHON as interpreter
This selects the right cpython abi version

However, maturin is still auto-selecting the wrong manylinux version
when we try and get it to detect that from the environment; a manylinux
version too new to be installed in, e.g. the conda python 39 build
environment, even though it was built there.  so keeping the
questionable `--compatibility linux` patch for today